+++ test/modules/auto_import.rb
@@ -1,0 +1,7 @@
+export :foo
+
+auto_import :BAR, './auto_import_bar'
+
+def foo
+  MODULE::BAR
+end