+++ test/modules/auto_import_bar.rb
@@ -1,0 +1,1 @@
+export_default 'bar'