+++ test/modules/auto_import_hash.rb
@@ -1,0 +1,8 @@
+export :M
+
+module M
+  auto_import(
+    BAR: './auto_import_bar',
+    BAZ: './auto_import_baz'
+  )
+end