+++ test/modules/auto_import_nested.rb
@@ -1,0 +1,5 @@
+export_default :M
+
+module M
+  auto_import :BAR, './auto_import_bar'
+end