+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Add auto_import feature for lazy loading of modules
+
 0.17 2018-11-22
 ---------------